Adding data to user-charity
