+++ src/components/progress/index.tsx
@@ -1,4 +1,5 @@
 import { createElement, FunctionComponent } from "react";
+import { humanNumber, humanTime } from "utils";
 import styles from "./style.scss";
 
 interface Props {
@@ -90,46 +91,3 @@ export const TimeProgress: FunctionComponent<Props> = (props) => {
         </svg>
     );
 };
-
-const SI_SYMBOL = ["", "k", "M", "G", "T", "P", "E"];
-
-function humanNumber(value: number): string {
-    // what tier? (determines SI symbol)
-    const tier = (Math.log10(value) / 3) | 0;
-
-    // if zero, we don't need a suffix
-    if (tier === 0) {
-        return value.toString();
-    }
-
-    // get suffix and determine scale
-    const suffix = SI_SYMBOL[tier];
-    const scale = Math.pow(10, tier * 3);
-
-    // scale the number
-    const scaled = value / scale;
-
-    // format number and add suffix
-    return scaled.toFixed(1) + suffix;
-}
-
-const TIME = ["S", "M", "H", "D"];
-function humanTime(seconds: number): string {
-    // Break points
-    const minuteBP = 60;
-    const hourBP = 3600;
-    const dayBP = 86400;
-
-    if (seconds > minuteBP && seconds < hourBP) {
-        const minutes = Math.floor(seconds / minuteBP);
-        return minutes.toFixed(0) + TIME[1];
-    } else if (seconds > hourBP && seconds < dayBP) {
-        const hours = Math.floor(seconds / hourBP);
-        return hours.toFixed(0) + TIME[2];
-    } else if (seconds > dayBP) {
-        const days = Math.floor(seconds / dayBP);
-        return days.toFixed(0) + TIME[3];
-    } else {
-        return seconds.toFixed(0) + TIME[0];
-    }
-}
