+++ src/components/user-charity/index.tsx
@@ -2,8 +2,11 @@ import Icon from "components/icon";
 import { createElement, FunctionComponent } from "react";
 import styles from "./style.scss";
 import Charity from "orm/charity";
+import { useUser } from "fb";
+import { get } from "lodash";
 import DonateABox from "components/donateabox";
 import LabelCircle from "components/label-circle";
+import { humanNumber, humanTime } from "utils";
 
 const defaultImg = () => {
     <DonateABox char={"D"} fontSize={"24"} />;
@@ -14,6 +17,15 @@ const UserCharityRank: FunctionComponent<{
     className?: string;
     charity?: Charity;
 }> = ({ rank, className, charity }) => {
+    const user = useUser();
+    const currentCharityHashes = charity?.shortName.toLowerCase() + "Hashes";
+    const currentCharityTime = charity?.shortName.toLowerCase() + "Time";
+
+    const userHashes: number = Number(get(user, currentCharityHashes));
+    const userTime: number = Number(get(user, currentCharityTime));
+    const formattedHashes: string = humanNumber(userHashes);
+    const formattedTime: string = humanTime(userTime);
+
     return (
         <div className={styles.container}>
             <div className={styles.rank}>
@@ -28,11 +40,17 @@ const UserCharityRank: FunctionComponent<{
                 </div>
                 <div className={styles.charityInfo}>
                     Total Hashes Donated
-                    <span className={styles.charityNumbers}> 123</span>
+                    <span className={styles.charityNumbers}>
+                        {" "}
+                        {formattedHashes}
+                    </span>
                 </div>
                 <div className={styles.charityInfo}>
                     Total Time Donated
-                    <span className={styles.charityNumbers}> 456</span>
+                    <span className={styles.charityNumbers}>
+                        {" "}
+                        {formattedTime}
+                    </span>
                 </div>
             </div>
         </div>
