+++ src/seed/charity.json
@@ -40,7 +40,7 @@
         ]
     },
     {
-        "shortName": "",
+        "shortName": "DBL",
         "longName": "donateABLE",
         "logo": "https://firebasestorage.googleapis.com/v0/b/donateable-42f3d.appspot.com/o/logo%2F3974b047-340b-47d5-bbd9-2e8c9c92a413?alt=media&token=451fbc1f-3fd3-4471-9390-6a1381a59623",
         "shortDesc": "donateABLE allows anyone with a computer to make a difference by donating unused computer CPU and turning it into real monetary value.",
