+++ src/utils.ts
@@ -494,3 +494,50 @@ export function secondsToString(seconds: number): string {
     const timeString = date.toISOString().substr(11, 8);
     return timeString;
 }
+
+const SI_SYMBOL = ["", "k", "M", "G", "T", "P", "E"];
+
+export function humanNumber(value: number): string {
+    // what tier? (determines SI symbol)
+    const tier = (Math.log10(value) / 3) | 0;
+
+    // if zero, we don't need a suffix
+    if (tier === 0) {
+        return value.toString();
+    }
+
+    // get suffix and determine scale
+    const suffix = SI_SYMBOL[tier];
+    const scale = Math.pow(10, tier * 3);
+
+    // scale the number
+    const scaled = value / scale;
+
+    // format number and add suffix
+    return scaled.toFixed(1) + suffix;
+}
+
+const TIME = ["S", "M", "H", "D"];
+
+export function humanTime(seconds: number): string {
+    // Break points
+    const minuteBP = 60;
+    const hourBP = 3600;
+    const dayBP = 86400;
+
+    if (seconds > minuteBP && seconds < hourBP) {
+        const minutes = Math.floor(seconds / minuteBP);
+        return minutes.toFixed(0) + TIME[1];
+    } else if (seconds > hourBP && seconds < dayBP) {
+        const hours = Math.floor(seconds / hourBP);
+        return hours.toFixed(0) + TIME[2];
+    } else if (seconds > dayBP) {
+        const days = Math.floor(seconds / dayBP);
+        return days.toFixed(0) + TIME[3];
+    } else {
+        if (seconds === 0) {
+            return seconds.toFixed(0);
+        }
+        return seconds.toFixed(0) + TIME[0];
+    }
+}
